+++ package.json
@@ -18,6 +18,7 @@
     "react": "17.0.1",
     "react-dom": "17.0.1",
     "react-native": "0.64.3",
+    "react-native-svg": "^12.3.0",
     "react-native-view-shot": "3.1.2",
     "react-native-web": "0.17.1",
     "styled-components": "^5.3.5"
